requirements.txt
python-dotenv >= 1.0.1
bcrypt >= 4.2.1
Flask[async] >= 3.1.0
Flask-SQLAlchemy >= 3.1.1
psycopg[binary,pool] >= 3.2.3
beautifulsoup4 >= 4.12.3
lxml >= 5.3.0
tqdm >= 4.67.1
orjson >= 3.10.12 # Faster than json
waitress >= 3.0.2
redis-om >= 0.3.3
requests >= 2.32.3
PyJWT >= 2.10.1
argon2-cffi >= 23.1.0